Direct dispatch test skeleton
